requirements.txt
requests>=2.31.0
ephem>=4.1.5
python-dotenv>=1.0.0
